Refactoring Inoculator and Sequencer machines
