+++ src/main/java/binnie/genetics/machine/Acclimatiser.java
@@ -20,6 +20,7 @@ import binnie.core.machines.power.ComponentProcessIndefinate;
 import binnie.core.machines.power.ErrorState;
 import binnie.genetics.core.GeneticsGUI;
 import binnie.genetics.core.GeneticsTexture;
+import binnie.genetics.machine.inoculator.Inoculator;
 import forestry.api.apiculture.EnumBeeChromosome;
 import forestry.api.genetics.AlleleManager;
 import forestry.api.genetics.EnumTolerance;
