+++ src/main/java/binnie/genetics/machine/inoculator/ValidatorIndividualInoculate.java
@@ -1,0 +1,24 @@
+package binnie.genetics.machine.inoculator;
+
+import binnie.Binnie;
+import binnie.core.machines.inventory.SlotValidator;
+import forestry.api.genetics.AlleleManager;
+import forestry.api.genetics.ISpeciesRoot;
+import net.minecraft.item.ItemStack;
+
+public class ValidatorIndividualInoculate extends SlotValidator {
+	public ValidatorIndividualInoculate() {
+		super(null);
+	}
+
+	@Override
+	public boolean isValid(ItemStack object) {
+		ISpeciesRoot root = AlleleManager.alleleRegistry.getSpeciesRoot(object);
+		return root != null && Binnie.Genetics.getSystem(root).isDNAManipulable(object);
+	}
+
+	@Override
+	public String getTooltip() {
+		return "Inoculable Individual";
+	}
+}
