+++ src/main/java/binnie/genetics/machine/sequencer/Sequencer.java
@@ -1,0 +1,15 @@
+package binnie.genetics.machine.sequencer;
+
+import binnie.core.resource.BinnieIcon;
+
+public class Sequencer {
+	public static final int[] SLOT_RESERVE = new int[]{1, 2, 3, 4};
+	public static final int SLOT_DYE_INDEX = 0;
+	public static final int SLOT_TARGET_INDEX = 5;
+	public static final int SLOT_DONE_INDEX = 6;
+
+	public static BinnieIcon fxSeqA;
+	public static BinnieIcon fxSeqG;
+	public static BinnieIcon fxSeqT;
+	public static BinnieIcon fxSeqC;
+}
